\name{bayesPop-internal}
\Rdversion{1.1}

\alias{KannistoAxBx}
\alias{StoPopProj}
\alias{do.pop.trajectories.plot}
\alias{get.age.labels}
\alias{get.country.inputs}
\alias{get.pop.traj.quantiles}
\alias{get.pop.trajectories}
\alias{get.pop.trajectories.multiple.age}
\alias{get.prediction.year.index}
\alias{get.traj.index}
\alias{is.saved.pi}
\alias{load.inputs}
\alias{modifiedLC}
\alias{read.bayesPop.file}
\alias{read.pop.file}
\alias{runKannisto}
\alias{migdepratio}
\alias{depratioM}
\alias{depratioF}
\alias{pasfr.glob.norms}

\title{
Internal Functions and datasets of bayesPop
}

\description{
These functions and datasets are not to be used directly by the user.
}

\keyword{internal}
