\name{get.countries.table}
\Rdversion{1.1}
\alias{get.countries.table}
\alias{get.countries.table.bayesPop.prediction}

\title{
Accessing Country Information
}
\description{
The function returns a data frame containing codes and names of all countries used in the prediction.
}
\usage{
\method{get.countries.table}{bayesPop.prediction}(object, \dots)
}
\arguments{
  \item{object}{Object of class \code{\link{bayesPop.prediction}}.}
  \item{\dots}{Not used.}
}
%\details{
%}
\value{
Data frame with columns \code{code} and \code{name}.
}
%\references{
%% ~put references to the literature/web site here ~
%}
\author{
Hana Sevcikova
}
%\note{
%%  ~~further notes~~
%}

%\seealso{
%}
%\examples{
%}

\keyword{ attribute }

