\name{MLTbx}
\alias{MLTbx}
\docType{data}
\title{
Datasets on Lee-Carter bx for Modeled Countries
}
\description{
Datasets with values of the Lee-Carter bx parameter for countries where mortality was obtained using model life tables.
}
\usage{
    data(MLTbx)
}
\format{
  A data frame }
%\details{
%}
\source{
Data provided by the United Nations Population Division.
}
%\references{
%%  ~~ possibly secondary sources and usages ~~
%}
\examples{
data(MLTbx)
str(MLTbx)
}
\keyword{datasets}
